+++ features/fixtures/laravel9/app/Http/Middleware/HandledMiddlewareEx.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Bugsnag\BugsnagLaravel\Facades\Bugsnag;
+use Closure;
+use Exception;
+
+class HandledMiddlewareEx
+{
+    public function handle($request, Closure $next)
+    {
+        Bugsnag::notifyException(new Exception('Handled middleware exception'));
+        return $next($request);
+    }
+}