+++ tests/Request/LaravelRequestTest.php
@@ -1,31 +1,32 @@
 <?php
 
+// stub out php_sapi_name in the Illumnate\Foundation namespace so we can change
+// its behaviour. Laravel 5.0 - 5.8 use this to determine if the app is running
+// in a console (6.0+ use an environment variable or the PHP_SAPI constant)
+
+namespace Illuminate\Foundation;
+
+use Bugsnag\BugsnagLaravel\Tests\Stubs\PhpSapiNameStub;
+
+function php_sapi_name()
+{
+    return PhpSapiNameStub::get();
+}
+
 namespace Bugsnag\BugsnagLaravel\Tests\Request;
 
 use Bugsnag\BugsnagLaravel\Request\LaravelRequest;
 use Bugsnag\BugsnagLaravel\Request\LaravelResolver;
+use Bugsnag\BugsnagLaravel\Tests\AbstractTestCase;
+use Bugsnag\BugsnagLaravel\Tests\Stubs\PhpSapiNameStub;
 use Bugsnag\Request\ConsoleRequest;
 use Bugsnag\Request\RequestInterface;
-use GrahamCampbell\TestBenchCore\MockeryTrait;
-use Illuminate\Foundation\Application;
-use Illuminate\Http\Request;
-use Mockery;
-use PHPUnit_Framework_TestCase as TestCase;
 
-class LaravelRequestTest extends TestCase
+class LaravelRequestTest extends AbstractTestCase
 {
-    use MockeryTrait;
-
     public function testCanResolveConsoleRequest()
     {
-        $request = Mockery::mock(Request::class);
-        $app = Mockery::mock(Application::class);
-
-        $app->shouldReceive('runningInConsole')->once()->andReturn(true);
-        $app->shouldReceive('make')->once()->with(Request::class)->andReturn($request);
-        $request->shouldReceive('server')->once()->with('argv', [])->andReturn('test mock console command');
-
-        $resolver = new LaravelResolver($app);
+        $resolver = new LaravelResolver($this->app);
         $result = $resolver->resolve();
 
         $this->assertInstanceOf(RequestInterface::class, $result);
@@ -34,14 +35,25 @@ class LaravelRequestTest extends TestCase
 
     public function testCanResolveLaravelRequest()
     {
-        $resolver = new LaravelResolver($app = Mockery::mock(Application::class));
+        try {
+            // different versions of Laravel use slightly different mechanisms
+            // to determine if the app is 'runningInConsole'
+            PhpSapiNameStub::set('not cli');
+            $this->setEnvironmentVariable('APP_RUNNING_IN_CONSOLE', false);
 
-        $app->shouldReceive('runningInConsole')->once()->andReturn(false);
-        $app->shouldReceive('make')->once()->with(Request::class)->andReturn($request = Mockery::mock(Request::class));
+            // At this point we already have a Laravel app loaded so the environment
+            // variable won't be picked up. Refreshing the application forces
+            // the config to re-load
+            $this->refreshApplication();
 
-        $request = $resolver->resolve();
+            $resolver = new LaravelResolver($this->app);
+            $result = $resolver->resolve();
 
-        $this->assertInstanceOf(RequestInterface::class, $request);
-        $this->assertInstanceOf(LaravelRequest::class, $request);
+            $this->assertInstanceOf(RequestInterface::class, $result);
+            $this->assertInstanceOf(LaravelRequest::class, $result);
+        } finally {
+            PhpSapiNameStub::reset();
+            $this->removeEnvironmentVariable('APP_RUNNING_IN_CONSOLE');
+        }
     }
 }