+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+## 2.23.0 (2022-02-09)
+
+### Enhancements
+
+* Allow installation on Laravel 9 projects
+  [jdanio](https://github.com/jdanio)
+  [#470](https://github.com/bugsnag/bugsnag-laravel/pull/470)
+  [#477](https://github.com/bugsnag/bugsnag-laravel/pull/477)
+* Allow installing Bugsnag PSR Logger v2. This adds support for PSR Log v3
+  [#471](https://github.com/bugsnag/bugsnag-laravel/pull/471)
+
 ## 2.22.2 (2021-09-06)
 
 ### Bug Fixes